refactor: move retry helper into fedimint_core
- use it in download client config
